#296 fix events() path
