+++ src/main/java/com/amihaiemil/docker/Docker.java
@@ -26,6 +26,8 @@
 package com.amihaiemil.docker;
 
 import java.io.IOException;
+import java.io.Reader;
+
 import org.apache.http.client.HttpClient;
 
 /**
@@ -42,7 +44,17 @@ public interface Docker {
      * @throws IOException If there's network problem.
      */
     boolean ping() throws IOException;
-    
+
+    /**
+     * Follow the events on the server in real time.
+     * @return The events {@link Reader}.
+     * @throws IOException If an I/O error occurs.
+     * @throws UnexpectedResponseException If the API responds with an
+     *  unexpected status.
+     */
+    Reader events()
+        throws IOException, UnexpectedResponseException;
+
     /**
      * Entry point for the Containers API.
      * @return Containers.
