+++ src/main/java/com/amihaiemil/docker/DockerSystem.java
@@ -26,7 +26,6 @@
 package com.amihaiemil.docker;
 
 import java.io.IOException;
-import java.io.Reader;
 
 /**
  * Docker System API entry point.
@@ -44,15 +43,5 @@ public interface DockerSystem {
      *  unexpected status.
      */
     DiskSpaceInfo diskUsage()
-        throws IOException, UnexpectedResponseException;
-
-    /**
-     * Follow the events on the server in real time.
-     * @return The events {@link Reader}.
-     * @throws IOException If an I/O error occurs.
-     * @throws UnexpectedResponseException If the API responds with an
-     *  unexpected status.
-     */
-    Reader events()
         throws IOException, UnexpectedResponseException;
 }
